A bunch of changes
The first time I tried this i fucked up... so here is the second attempt.

Tried to fix the window canvas size, but it needs to be custom 16:9 I think. Currently 16:10.
