+++ Pong/Assets/Prefab/Ball.prefab.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
 guid: 4526f4845a0766740925c9e6d951a6bd
-timeCreated: 1490755645
-licenseType: Free
 NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
