+++ Pong/Assets/Prefab/RacketRight.prefab.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
 guid: febcad35c51de3b449589743b789f87d
-timeCreated: 1490755684
-licenseType: Free
 NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
