+++ Pong/Assets/Scripts/UpdateScore.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d1dc26c4c19aaab49adbd15dc3bfccbf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
